Patient Views: conditionally render links/button to other actions based upon the user's permissions

+++ resources/views/patients/index.blade.php
@@ -9,13 +9,15 @@
             {!! link_to_with_icon('fas fa-arrow-left fa-2x', 'home', [], 'Back to Home', 'has-text-info') !!}
         </div>
 
-        <div class="level-right">
+        @can ('create', \App\Patient::class)
+            <div class="level-right">
 
-            <div class="level-item">
-                {!! link_to_with_icon('fas fa-plus fa-2x', 'patients.create', [], 'Create a new Patient', 'has-text-success') !!}
+                <div class="level-item">
+                    {!! link_to_with_icon('fas fa-plus fa-2x', 'patients.create', [], 'Create a new Patient', 'has-text-success') !!}
+                </div>
+
             </div>
-
-        </div>
+        @endcan
 
     </div>
 
@@ -37,8 +39,14 @@
                     <th>Address</th>
                     <th>Phone</th>
                     <th>Details</th>
-                    <th>Edit</th>
-                    <th>Delete</th>
+
+                    @can ('update', \App\Patient::class)
+                        <th>Edit</th>
+                    @endcan
+
+                    @can ('delete', \App\Patient::class)
+                        <th>Delete</th>
+                    @endcan
                 </tr>
             </thead>
 
@@ -55,8 +63,14 @@
                         <td>{{ $patient->address }}</td>
                         <td>{{ $patient->phone }}</td>
                         <td>{!! link_to_with_icon('fas fa-info-circle fa-2x', 'patients.show', $patient->id, '', 'has-text-info') !!}</td>
-                        <td>{!! link_to_with_icon('fas fa-edit fa-2x', 'patients.edit', $patient, '', 'has-text-dark') !!}</td>
-                        <td>{!! delete_link_with_icon('fas fa-trash-alt fa-2x', 'patients.destroy', $patient) !!}</td>
+
+                        @can ('update', $patient)
+                            <td>{!! link_to_with_icon('fas fa-edit fa-2x', 'patients.edit', $patient, '', 'has-text-dark') !!}</td>
+                        @endcan
+
+                        @can ('delete', $patient)
+                            <td>{!! delete_link_with_icon('fas fa-trash-alt fa-2x', 'patients.destroy', $patient) !!}</td>
+                        @endcan
                     </tr>
                 @endforeach
             </tbody>
